cd: added .ecm file format support #41
Due to ECM format design it is impossible to access data in random order.
Whole file needs to be processed first to build LUT table for offsets of sectors in .ecm file.
This implementation loads the whole processed file to RAM which will result in up to 800MB allocated by emulator when loading .ecm file.
Loading might freeze emulator for few seconds depending on drive speed - loader is single-threaded for now.